Separate SCF and StaticLogic to Calyx conversions.
