+++ lib/Conversion/CMakeLists.txt
@@ -8,5 +8,6 @@ add_subdirectory(HandshakeToHW)
 add_subdirectory(LLHDToLLVM)
 add_subdirectory(MooreToCore)
 add_subdirectory(SCFToCalyx)
+add_subdirectory(StaticLogicToCalyx)
 add_subdirectory(StandardToHandshake)
 add_subdirectory(StandardToStaticLogic)
