+++ test/Conversion/StaticLogicToCalyx/convert_pipeline.mlir
@@ -1,4 +1,4 @@
-// RUN: circt-opt %s -lower-scf-to-calyx -split-input-file | FileCheck %s
+// RUN: circt-opt %s -lower-static-logic-to-calyx -split-input-file | FileCheck %s
 
 // CHECK:     calyx.program "minimal"
 // CHECK:       calyx.component @minimal
